💥refactor: CoreDataManager을 제너릭타입 활용으로 변경
