+++ Diary/Application/AppDelegate.swift
@@ -5,7 +5,6 @@
 // 
 
 import UIKit
-import CoreData
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
